+++ modules/react/tabs/lib/menu/MenuItem.tsx
@@ -11,13 +11,13 @@ import {
 } from '@workday/canvas-kit-react/common';
 import {SystemIcon} from '@workday/canvas-kit-react/icon';
 import {OverflowTooltip} from '@workday/canvas-kit-react/tooltip';
+import {Box} from '@workday/canvas-kit-labs-react/common';
 
 import {useListRegisterItem} from '../list';
 import {useRovingFocus} from '../cursor';
 import {MenuModel} from './useMenuModel';
 import {MenuModelContext} from './Menu';
 import {useSelectionItem} from '../selection';
-import {Box} from '@workday/canvas-kit-labs-react/common';
 
 export interface MenuItemProps<T = unknown> {
   /**
