+++ modules/react/tabs/lib/menu/MenuPopper.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 
 import {createComponent, ExtractProps, useModelContext} from '@workday/canvas-kit-react/common';
-
 import {usePopupPopper, Popper} from '@workday/canvas-kit-react/popup';
 
 import {MenuModel} from './useMenuModel';
