+++ modules/react/tabs/lib/overflow/index.ts
@@ -1,3 +1,4 @@
 export * from './useOverflowModel';
-export * from './useMeasureContainer';
+export * from './useOverflowMeasureItem';
+export * from './useOverflowMeasureContainer';
 export * from './useOverflowTarget';
